requirements.txt
faiss-cpu
numpy
sentence-transformers
requests
PyMuPDF